Benchmark: Move code out of Timer.h
